Touchups:
Removed old TODO comments.
Deleted 'BasePlayer' module & the Somna implementation using it.
Modified Dice to return a float - originally returned an int, which would be treated as a string and require (num) conversion.
Interpreter: removed the 'silent scene' check as it was misplaced and would never work - it required knowledge of the results of interpretation before interpretation took place.

+++ CYA Adventure Game Engine/Modules/Dice.cs
@@ -2,7 +2,7 @@
 {
     public class Dice : IModule
     {
-        public static int Roll(string rollType)
+        public static float Roll(string rollType)
         {
             var rand = new Random();
             if (!(rollType.Contains('d')))
@@ -21,7 +21,7 @@
             {
                 total += rand.Next(1, diceSides + 1);
             }
-            return total;
+            return (float)total;
         }
     }
 }
